enhance: Remove merge logic sharing between Entity and Collection
