+++ packages/endpoint/src/schemas/Collection.ts
@@ -3,10 +3,10 @@ import {
   INormalizeDelegate,
   PolymorphicInterface,
   IQueryDelegate,
+  Mergeable,
 } from '../interface.js';
 import { Values, Array as ArraySchema } from '../schema.js';
 import type { DefaultArgs } from '../schemaTypes.js';
-import { Mergeable } from './Mergeable.js';
 
 const pushMerge = (existing: any, incoming: any) => {
   return [...existing, ...incoming];
@@ -28,7 +28,8 @@ export default class CollectionSchema<
   S extends PolymorphicInterface = any,
   Args extends any[] = DefaultArgs,
   Parent = any,
-> extends Mergeable {
+> implements Mergeable
+{
   declare protected nestKey: (parent: any, key: string) => Record<string, any>;
 
   declare protected argsKey?: (...args: any) => Record<string, any>;
@@ -76,7 +77,6 @@ export default class CollectionSchema<
   }
 
   constructor(schema: S, options?: CollectionOptions<Args, Parent>) {
-    super();
     this.schema =
       Array.isArray(schema) ? (new ArraySchema(schema[0]) as any) : schema;
     if (!options) {
@@ -175,6 +175,44 @@ export default class CollectionSchema<
   // always replace
   merge(existing: any, incoming: any) {
     return incoming;
+  }
+
+  shouldReorder(
+    existingMeta: { date: number; fetchedAt: number },
+    incomingMeta: { date: number; fetchedAt: number },
+    existing: any,
+    incoming: any,
+  ) {
+    return incomingMeta.fetchedAt < existingMeta.fetchedAt;
+  }
+
+  mergeWithStore(
+    existingMeta: {
+      date: number;
+      fetchedAt: number;
+    },
+    incomingMeta: { date: number; fetchedAt: number },
+    existing: any,
+    incoming: any,
+  ) {
+    return this.shouldReorder(existingMeta, incomingMeta, existing, incoming) ?
+        this.merge(incoming, existing)
+      : this.merge(existing, incoming);
+  }
+
+  mergeMetaWithStore(
+    existingMeta: {
+      fetchedAt: number;
+      date: number;
+      expiresAt: number;
+    },
+    incomingMeta: { fetchedAt: number; date: number; expiresAt: number },
+    existing: any,
+    incoming: any,
+  ) {
+    return this.shouldReorder(existingMeta, incomingMeta, existing, incoming) ?
+        existingMeta
+      : incomingMeta;
   }
 
   // >>>>>>>>>>>>>>DENORMALIZE<<<<<<<<<<<<<<
