+++ packages/normalizr/src/normalize/NormalizeDelegate.ts
@@ -84,7 +84,7 @@ export class NormalizeDelegate
 
   /** Updates an entity using merge lifecycles when it has previously been set */
   mergeEntity(
-    schema: Mergeable & { key: string; indexes?: any },
+    schema: Mergeable & { indexes?: any },
     pk: string,
     incomingEntity: any,
   ) {
